+++ public/css/tricks.css
@@ -502,7 +502,6 @@ img {
 
 .btn-info {
   margin-top: 30px;
-  text-align: right;
 }
 
 label {
