+++ public/js/scripts.js
@@ -44,10 +44,15 @@
     });
 
     // Activate scrollspy to add active class to navbar items on scroll
-    $('body').scrollspy({
-        target: '#mainNav',
-        offset: 80
-    });
+    // $('body').scrollSpy({
+    //     target: '#mainNav',
+    //     offset: 80
+    // });
+
+    // var scrollSpy = new bootstrap.ScrollSpy(document.body, {
+    //     target: '#mainNav',
+    //     offset: 80
+    // })
 
     // Collapse Navbar  // not used....
     var navbarCollapse = function () {
