+++ src/Controller/HomeController.php
@@ -10,13 +10,13 @@ use Symfony\Component\Routing\Annotation\Route;
 /**
  * class MainControler : homepage
  */
-class MainControlerController extends AbstractController
+class HomeController extends AbstractController
 {
     #[Route('/', name: 'app_home')]
     public function index(TricksRepository $TricksRepository): Response
     {
         return $this->render('main_controler/index.html.twig', [
-            'controller_name' => 'MainControlerController',
+            'controller_name' => 'HomeController',
             'tricks' => $TricksRepository->findBy(['active' => true], ['created_at' => 'asc'])
         ]);
     }
