+++ templates/main_controler/index.html.twig
@@ -14,7 +14,7 @@
 
         <!-- go down to the tricks -->
         <div>
-            <a class="js-scroll-trigger d-block text-center text-white rounded positionArrow" href="#tricks">
+            <a class="d-block text-center text-white rounded positionArrow" href="#tricks">
                 <i class=" bi bi-arrow-down-circle-fill fa-3x chevron-down"></i>
                 <br>
                 <br>
