+++ templates/tricks/details.html.twig
@@ -45,7 +45,7 @@ Snowtricks - {{ routeTitre }} - {{ trick.title }}
 
                     {# main picture tools#}
                     <span>
-
+                        {# {{ trick.picture }} #}
                         <div class="form-group pencilMainPictureAction displayNone">
                             {{ form_row(formUpdateTrick.picture, {
                             label: 'Change the main picture',
@@ -75,10 +75,10 @@ Snowtricks - {{ routeTitre }} - {{ trick.title }}
                             {% if is_granted('IS_AUTHENTICATED_REMEMBERED') and app.request.attributes.get('_route') ==
                             "tricks_modifications" %}
 
-                            <a href="#" class="float-right ml-1">
+                            <a href="#title-trick" class="float-right ml-1">
                                 <i class="bi bi-pencil-fill pencilTitleAction iconTricks"></i>
                             </a>
-
+                            <span id="title-trick"></span>
                             <div class="form-group titleAction displayNone">
                                 <br>
                                 {{ form_row(formUpdateTrick.title, {
@@ -86,25 +86,15 @@ Snowtricks - {{ routeTitre }} - {{ trick.title }}
                                 attr : {'class': 'form-control smallForForm'}
                                 }) }}
                             </div>
-
-                            {# <a href="/tricks/modification/{{ trick.slug }}" class="float-right ml-1"> #}
-
-                                {% endif %}
+                            {% endif %}
                         </h1>
                     </div>
                 </div>
-
 
                 {####### additionnals pictures #########}
                 <div class="row justify-content-center">
 
                     {% for additionnalPicture in additionnalPictures %}
-
-                    {#
-                    <pre>
-                    {{ additionnalPicture.picure }}
-                    </pre> #}
-
                     <!-- additionnalPictures Items-->
                     <div class="col-md-2 col-lg-2 mb-5 item">
                         <div class="mx-auto">
@@ -171,18 +161,20 @@ Snowtricks - {{ routeTitre }} - {{ trick.title }}
 
                 </div>
 
-                {# additionnal picture here #}
+                {# additionnal picture #}
+                {% if is_granted('IS_AUTHENTICATED_REMEMBERED') and
+                app.request.attributes.get('_route') == "tricks_modifications" %}
                 <div class="container text-center">
                     <div class="row">
                         <div class="form-group">
                             {{ form_row(formUpdateTrick.pictures, {
-                            label: 'Add more additionnal pictues',
+                            label: 'Add more additionnal pictures',
                             attr : {'class': 'form-control'}
                             }) }}
                         </div>
                     </div>
                 </div>
-
+                {% endif %}
 
                 {####### information #########}
                 <div class="card-body">
@@ -237,17 +229,17 @@ Snowtricks - {{ routeTitre }} - {{ trick.title }}
                         label: 'Groups',
                         attr : {'class': 'form-control'}
                         }) }}
-                        {% else %}
-                        {# {{ trick.category }} #}
                         {% endif %}
                     </div>
 
-                    <div class="col-2 justify-content-end">
+                    <div class="col-2 float-right ml-1">
                         {##### Button delete #######}
                         {{ include('tricks/tricks_delete_form.html.twig') }}
                     </div>
-                    <div class="col-2 justify-content-end">
+                    {##### Button save #######}
+                    <div class="col-2 float-right ml-1">
                         <button class=" btn btn-info" type="submit">save</button>
+                        {# <button class="btn">{{ button_label|default('Save') }}</button> #}
                     </div>
                 </div>
             </div>
