+++ templates/tricks/tricks_delete_form.html.twig
@@ -3,3 +3,4 @@
     <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ trick.id) }}">
     <button class="btn btn-danger">Delete</button>
 </form>
+{# <small><i class="bi bi-arrow-right-short white"></i></small> #}
